extracted scene tilemap build mechanics into their own class
